refactor: split user message into separate component files
